requirements.txt
PyPDF2
Pillow
gunicorn
flask>=2.2.0
werkzeug==2.0.3